Move validRegex to Commands object (#276)
