+++ kernel/src/test/scala/me/rexim/morganey/MonadSpec.scala
@@ -1,10 +1,9 @@
-package me.rexim.morganey.util
+package me.rexim.morganey
 
 import me.rexim.morganey.monad._
 import org.scalatest._
 
-class UtilSpec extends FlatSpec with Matchers {
-
+class MonadSpec extends FlatSpec with Matchers {
   "The sequence function" should "swap the monadic context in" in {
     val numbers  = List(1 to 10: _*)
     val someNums = numbers.map(Option(_))
@@ -16,15 +15,4 @@ class UtilSpec extends FlatSpec with Matchers {
     sequence(List(Some(1))) should be (Some(List(1)))
     sequence(someNums) should be (Some(numbers))
   }
-
-  "The validRegex function" should "give back a function to match strings, if a valid regex was given" in {
-    val matcher1 = validRegex("[a-zA-Z]+")
-    matcher1                      shouldBe a[Some[_]]
-    matcher1 exists (_("fooBar")) should be (true)
-
-    val matcher2 = validRegex("[*")
-    matcher2                      should be (None)
-  }
-
-
 }
